requirements.txt
Flask>=2.0
psycopg2-binary>=2.8
azure-servicebus>=7.0.0
authlib>=1.0.0